hw3_vit_caption/requirements.txt

certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.10.0
filelock==3.4.0
gdown==4.2.0
huggingface-hub==0.1.2
idna==3.3
imageio==2.9.0
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.3
pandas==1.1.3
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-pretrained-vit==0.0.7
pytz==2020.1
PyWavelets==1.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.6.2
six==1.16.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
transformers==4.12.5
typing_extensions==4.0.0
urllib3==1.26.7
